nnnn

+++ backend/src/controllers/books.controller.js
@@ -82,32 +82,34 @@ const getAllBooks = asyncHandler(async (req, res) => {
       });
       
 
-    const deleteBook = asyncHandler(async (req, res) => {
-        const bookId = req.params.bookId;
-        try {
-          const bookDeleted = await Books.findOneAndDelete({ _id: bookId });
-          if (!bookDeleted) {
-            throw new ApiError(400, "Book not deleted");
-          }
-      
-        //   const filePath = path.join(__dirname, '../../public', bookDeleted.bookPath);
-        //   if (fs.existsSync(filePath)) {
-        //     fs.unlinkSync(filePath);
-        //   } else {
-        //     console.log(`File not found: ${filePath}`);
-        //   }
-
-          const removedBook = fs.unlinkSync(bookDeleted.bookPath)
-
-          if(!removedBook){
-            throw new ApiError(402, "Book is not removed from temp")
-          }
+      const deleteBook = asyncHandler(async (req, res) => {
+        
+            const bookId = req.params.bookId;
+          
+            try {
+              // Find the department to get the thumbnail path
+              const bookToDelete = await Books.findOne({ _id: bookId });
+          
+              if (!bookToDelete) {
+                throw new ApiError(400, "Book not available");              }
+          
+              // Remove the department from the database
+              await Books.findOneAndDelete({ _id: bookId });
+          
+              // Delete the thumbnail image from the public folder
+              const bookPath = path.join('public', bookToDelete.bookPath.replace('/public/', ''));
+              
+              if (fs.existsSync(bookPath)) {
+                fs.unlinkSync(bookPath);
+              }
+          
       
           return res.status(200).json(new ApiResponse(200, {}, "Book deleted successfully"));
         } catch (error) {
           return res.status(500).json(new ApiResponse(500, {}, error.message || "Something went wrong"));
         }
       });
+    
     
 
       const publishedBooks = asyncHandler( async (req, res) => {
